builtins: implement ST_S2Covering with optional params
Also refactored var parsing to paramparse to avoid the builtins <-> sql
circular dependency.

Release note (sql change): implement ST_S2Covering with optional index
parameters.
