+++ pkg/sql/vars.go
@@ -25,6 +25,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/catconstants"
 	"github.com/cockroachdb/cockroach/pkg/sql/delegate"
 	"github.com/cockroachdb/cockroach/pkg/sql/lex"
+	"github.com/cockroachdb/cockroach/pkg/sql/paramparse"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgnotice"
@@ -94,26 +95,6 @@ func formatBoolAsPostgresSetting(b bool) string {
 	return "off"
 }
 
-func parseBoolVar(varName, val string) (bool, error) {
-	val = strings.ToLower(val)
-	switch val {
-	case "on":
-		return true, nil
-	case "off":
-		return false, nil
-	case "yes":
-		return true, nil
-	case "no":
-		return false, nil
-	}
-	b, err := strconv.ParseBool(val)
-	if err != nil {
-		return false, pgerror.Newf(pgcode.InvalidParameterValue,
-			"parameter \"%s\" requires a Boolean value", varName)
-	}
-	return b, nil
-}
-
 // makeDummyBooleanSessionVar generates a sessionVar for a bool session setting.
 // These functions allow the setting to be changed, but whose values are not used.
 // They are logged to telemetry and output a notice that these are unused.
@@ -126,7 +107,7 @@ func makeDummyBooleanSessionVar(
 	return sessionVar{
 		GetStringVal: makePostgresBoolGetStringValFn(name),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar(name, s)
+			b, err := paramparse.ParseBoolVar(name, s)
 			if err != nil {
 				return err
 			}
@@ -376,7 +357,7 @@ var varGen = map[string]sessionVar{
 	`default_transaction_read_only`: {
 		GetStringVal: makePostgresBoolGetStringValFn("default_transaction_read_only"),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("default_transaction_read_only", s)
+			b, err := paramparse.ParseBoolVar("default_transaction_read_only", s)
 			if err != nil {
 				return err
 			}
@@ -431,7 +412,7 @@ var varGen = map[string]sessionVar{
 	`disable_partially_distributed_plans`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`disable_partially_distributed_plans`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("disable_partially_distributed_plans", s)
+			b, err := paramparse.ParseBoolVar("disable_partially_distributed_plans", s)
 			if err != nil {
 				return err
 			}
@@ -450,7 +431,7 @@ var varGen = map[string]sessionVar{
 	`enable_zigzag_join`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`enable_zigzag_join`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("enable_zigzag_join", s)
+			b, err := paramparse.ParseBoolVar("enable_zigzag_join", s)
 			if err != nil {
 				return err
 			}
@@ -492,7 +473,7 @@ var varGen = map[string]sessionVar{
 	`require_explicit_primary_keys`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`require_explicit_primary_keys`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("require_explicit_primary_key", s)
+			b, err := paramparse.ParseBoolVar("require_explicit_primary_key", s)
 			if err != nil {
 				return err
 			}
@@ -594,7 +575,7 @@ var varGen = map[string]sessionVar{
 	`optimizer_use_histograms`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`optimizer_use_histograms`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("optimizer_use_histograms", s)
+			b, err := paramparse.ParseBoolVar("optimizer_use_histograms", s)
 			if err != nil {
 				return err
 			}
@@ -613,7 +594,7 @@ var varGen = map[string]sessionVar{
 	`optimizer_use_multicol_stats`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`optimizer_use_multicol_stats`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("optimizer_use_multicol_stats", s)
+			b, err := paramparse.ParseBoolVar("optimizer_use_multicol_stats", s)
 			if err != nil {
 				return err
 			}
@@ -632,7 +613,7 @@ var varGen = map[string]sessionVar{
 	`enable_implicit_select_for_update`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`enable_implicit_select_for_update`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("enabled_implicit_select_for_update", s)
+			b, err := paramparse.ParseBoolVar("enabled_implicit_select_for_update", s)
 			if err != nil {
 				return err
 			}
@@ -651,7 +632,7 @@ var varGen = map[string]sessionVar{
 	`enable_insert_fast_path`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`enable_insert_fast_path`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("enable_insert_fast_path", s)
+			b, err := paramparse.ParseBoolVar("enable_insert_fast_path", s)
 			if err != nil {
 				return err
 			}
@@ -670,7 +651,7 @@ var varGen = map[string]sessionVar{
 	`enable_interleaved_joins`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`enable_interleaved_joins`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("enable_interleaved_joins", s)
+			b, err := paramparse.ParseBoolVar("enable_interleaved_joins", s)
 			if err != nil {
 				return err
 			}
@@ -739,7 +720,7 @@ var varGen = map[string]sessionVar{
 		},
 		GetStringVal: makePostgresBoolGetStringValFn("force_savepoint_restart"),
 		Set: func(_ context.Context, m *sessionDataMutator, val string) error {
-			b, err := parseBoolVar("force_savepoint_restart", val)
+			b, err := paramparse.ParseBoolVar("force_savepoint_restart", val)
 			if err != nil {
 				return err
 			}
@@ -801,7 +782,7 @@ var varGen = map[string]sessionVar{
 		},
 		GetStringVal: makePostgresBoolGetStringValFn("sql_safe_updates"),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("sql_safe_updates", s)
+			b, err := paramparse.ParseBoolVar("sql_safe_updates", s)
 			if err != nil {
 				return err
 			}
@@ -818,7 +799,7 @@ var varGen = map[string]sessionVar{
 		},
 		GetStringVal: makePostgresBoolGetStringValFn("prefer_lookup_joins_for_fks"),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("prefer_lookup_joins_for_fks", s)
+			b, err := paramparse.ParseBoolVar("prefer_lookup_joins_for_fks", s)
 			if err != nil {
 				return err
 			}
@@ -839,7 +820,7 @@ var varGen = map[string]sessionVar{
 			comma := ""
 			var buf bytes.Buffer
 			for _, v := range values {
-				s, err := datumAsString(&evalCtx.EvalContext, "search_path", v)
+				s, err := paramparse.DatumAsString(&evalCtx.EvalContext, "search_path", v)
 				if err != nil {
 					return "", err
 				}
@@ -1013,7 +994,7 @@ var varGen = map[string]sessionVar{
 	`transaction_read_only`: {
 		GetStringVal: makePostgresBoolGetStringValFn("transaction_read_only"),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("transaction_read_only", s)
+			b, err := paramparse.ParseBoolVar("transaction_read_only", s)
 			if err != nil {
 				return err
 			}
@@ -1052,7 +1033,7 @@ var varGen = map[string]sessionVar{
 			return formatBoolAsPostgresSetting(evalCtx.SessionData.AllowPrepareAsOptPlan)
 		},
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("allow_prepare_as_opt_plan", s)
+			b, err := paramparse.ParseBoolVar("allow_prepare_as_opt_plan", s)
 			if err != nil {
 				return err
 			}
@@ -1079,7 +1060,7 @@ var varGen = map[string]sessionVar{
 	`experimental_enable_temp_tables`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`experimental_enable_temp_tables`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("experimental_enable_temp_tables", s)
+			b, err := paramparse.ParseBoolVar("experimental_enable_temp_tables", s)
 			if err != nil {
 				return err
 			}
@@ -1098,7 +1079,7 @@ var varGen = map[string]sessionVar{
 	`experimental_enable_hash_sharded_indexes`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`experimental_enable_hash_sharded_indexes`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("experimental_enable_hash_sharded_indexes", s)
+			b, err := paramparse.ParseBoolVar("experimental_enable_hash_sharded_indexes", s)
 			if err != nil {
 				return err
 			}
@@ -1116,7 +1097,7 @@ var varGen = map[string]sessionVar{
 	`disallow_full_table_scans`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`disallow_full_table_scan`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar(`disallow_full_table_scans`, s)
+			b, err := paramparse.ParseBoolVar(`disallow_full_table_scans`, s)
 			if err != nil {
 				return err
 			}
@@ -1135,7 +1116,7 @@ var varGen = map[string]sessionVar{
 	`enable_experimental_alter_column_type_general`: {
 		GetStringVal: makePostgresBoolGetStringValFn(`enable_experimental_alter_column_type_general`),
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar("enable_experimental_alter_column_type_general", s)
+			b, err := paramparse.ParseBoolVar("enable_experimental_alter_column_type_general", s)
 			if err != nil {
 				return err
 			}
@@ -1188,7 +1169,7 @@ func makePostgresBoolGetStringValFn(varName string) getStringValFn {
 		if s, ok := val.(*tree.DString); ok {
 			return string(*s), nil
 		}
-		s, err := getSingleBool(varName, val)
+		s, err := paramparse.GetSingleBool(varName, val)
 		if err != nil {
 			return "", err
 		}
@@ -1237,7 +1218,7 @@ func makeCompatBoolVar(varName string, displayValue, anyValAllowed bool) session
 	return sessionVar{
 		Get: func(_ *extendedEvalContext) string { return displayValStr },
 		Set: func(_ context.Context, m *sessionDataMutator, s string) error {
-			b, err := parseBoolVar(varName, s)
+			b, err := paramparse.ParseBoolVar(varName, s)
 			if err != nil {
 				return err
 			}
@@ -1311,20 +1292,6 @@ func IsSessionVariableConfigurable(varName string) (exists, configurable bool) {
 }
 
 var varNames []string
-
-// getSingleBool returns the boolean if the input Datum is a DBool,
-// and returns a detailed error message if not.
-func getSingleBool(name string, val tree.Datum) (*tree.DBool, error) {
-	b, ok := val.(*tree.DBool)
-	if !ok {
-		err := pgerror.Newf(pgcode.InvalidParameterValue,
-			"parameter %q requires a Boolean value", name)
-		err = errors.WithDetailf(err,
-			"%s is a %s", val, errors.Safe(val.ResolvedType()))
-		return nil, err
-	}
-	return b, nil
-}
 
 func getSessionVar(name string, missingOk bool) (bool, sessionVar, error) {
 	if _, ok := UnsupportedVars[name]; ok {
